auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["mmskv <max.mishuk2014@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
authlib = "^1.2.0"
flask = "^2.2.3"
requests = "^2.28.2"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.16.2"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
